dashboard/requirements.txt
fastapi
uvicorn
paho-mqtt
fastapi_mqtt
haversine